chore: checking alternative lib
