+++ packages/respect-core/package.json
@@ -45,6 +45,7 @@
     "jest-matcher-utils": "^29.3.1",
     "json-pointer": "^0.6.2",
     "jsonpath-plus": "^10.0.6",
+    "jsonpath-rfc9535": "^1.3.0",
     "openapi-sampler": "^1.6.1",
     "outdent": "^0.8.0"
   },
